+++ src/console/controllers/SentNotificationsController.php
@@ -58,11 +58,11 @@ class SentNotificationsController extends Controller
         if ($this->formHandle !== null) {
             $formHandle = explode(',', $this->formHandle);
 
-            $formIds = Form::find()->handle($formHandle)->ids();
+            $formIds = Form::find()->handle($formHandle)->site('*')->unique()->ids();
         }
 
         if ($this->all) {
-            $formIds = Form::find()->ids();
+            $formIds = Form::find()->site('*')->unique()->ids();
         }
 
         if (!$this->formId && !$this->formHandle && !$this->all) {